Extract AutoYaST partitions/stray blk device planning
